Updating for indexController
