--- src/main/java/com/IndexController_inSpringBoot.java
@@ -1,72 +1,0 @@
-package com;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import com.tia102g5.article.model.Article;
-import com.tia102g5.article.model.ArticleService;
-import com.tia102g5.articleImg.model.ArticleImgService;
-import com.tia102g5.board.model.Board;
-import com.tia102g5.board.model.BoardService;
-
-
-
-//@PropertySource("classpath:application.properties") // 於https://start.spring.io建立Spring Boot專案時, application.properties文件預設已經放在我們的src/main/resources 目錄中，它會被自動檢測到
-@Controller
-public class IndexController_inSpringBoot {
-	
-	// @Autowired (●自動裝配)(Spring ORM 課程)
-	// 目前自動裝配了EmpService --> 供第66使用
-
-	
-    // inject(注入資料) via application.properties
-    @Value("${welcome.message}")
-    private String message;
-	
-    private List<String> myList = Arrays.asList("Spring Boot Quickstart 官網 : https://start.spring.io", "IDE 開發工具", "直接使用(匯入)官方的 Maven Spring-Boot-demo Project + pom.xml", "直接使用官方現成的 @SpringBootApplication + SpringBootServletInitializer 組態檔", "依賴注入(DI) HikariDataSource (官方建議的連線池)", "Thymeleaf", "Java WebApp (<font color=red>快速完成 Spring Boot Web MVC</font>)");
-    @GetMapping("/")
-    public String index(Model model) {
-    	model.addAttribute("message", message);
-        model.addAttribute("myList", myList);
-        return "index"; //view
-    }
-    
-    // http://......../hello?name=peter1
-    @GetMapping("/hello")
-    public String indexWithParam(
-            @RequestParam(name = "name", required = false, defaultValue = "") String name, Model model) {
-        model.addAttribute("message", name);
-        return "index"; //view
-    }
-    
-    @GetMapping("/header")
-    public String getHeader() {
-        return "header";
-    }
-    
-    @GetMapping("/footer")
-    public String getFooter() {
-        return "footer";
-    }
-    
-    //導向後台主頁
-    @GetMapping("/backEndPartner")
-    public String getPartner() {
-        return "/back-end-partner/partner";
-    }
-    
-    //後臺主頁側邊欄，用於 partner.js 第二行
-    @GetMapping("/partnerSidebar")
-    public String getPartnerSidebar() {
-        return "back-end-partner/partner_sidebar"; 
-    }
-
-}
